Python/voice-assistant/requirements.txt
SpeechRecognition
pyttsx3
pyaudio
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

